fix circular import for _io_repeat
