+++ PFERD/crawl/ilias/kit_ilias_web_crawler.py
@@ -7,8 +7,9 @@ from bs4 import BeautifulSoup
 from ...auth import Authenticator, TfaAuthenticator
 from ...config import Config
 from ...logging import log
-from ...utils import _iorepeat, soupify
+from ...utils import soupify
 from ..crawler import CrawlError, CrawlWarning
+from .async_helper import _iorepeat
 from .ilias_web_crawler import IliasWebCrawler, IliasWebCrawlerSection
 
 TargetType = Union[str, int]
